Thêm outbox pattern có sẵn của masstransit, test thử đã lưu vào db, bg job consume nó, message ở trên queue, thêm customOpenApi cho jwt authen

+++ src/Services/Orders/Ecommerce.Services.Orders.Contracts/Events/OrderCreatedEvent.cs
@@ -1,10 +1,12 @@
 using BuildingBlocks.Messaging.Abstractions;
+using BuildingBlocks.Shared.InfrastructureInterfaces.Messaging;
 
 namespace Ecommerce.Services.Orders.Contracts.Events;
 
 public record OrderCreatedEvent : IIntegrationEvent
+
 {
     public Guid Id { get; set; }
     public DateTime CreatedAt { get; set; }
-    public int CustomerId { get; set; }
+    public long CustomerId { get; set; }
 }
